Added GStreamer support using GStreamer-Playbin by the Dissent project.
svn path=/trunk/; revision=866

+++ src/GnomeSubtitles/Core/Util.cs
@@ -44,7 +44,7 @@ public class Util {
 			":" + time.Seconds.ToString("00;00") + "." + time.Milliseconds.ToString("000;000");
 	}
 	
-	public static string SecondsToTimeText (float seconds) {
+	public static string SecondsToTimeText (double seconds) {
 		return TimeSpanToText(TimeSpan.FromSeconds(seconds));
 	}
